+++ util/dataset.go
@@ -12,13 +12,16 @@ import (
 )
 
 type DataSet struct {
+	ID             string                   `config:"id" json:"id"`
 	Title          string                   `config:"title" json:"title" validate:"required"`
-	Name           string                   `config:"name" json:"name"`
 	Release        string                   `config:"release" json:"release"`
 	Type           string                   `config:"type" json:"type" validate:"required"`
 	IngestPipeline string                   `config:"ingest_pipeline,omitempty" config:"ingest_pipeline" json:"ingest_pipeline,omitempty"`
 	Vars           []map[string]interface{} `config:"vars" json:"vars,omitempty"`
 	Package        string                   `json:"package"`
+
+	// Generated fields
+	Path string `json:"path"`
 }
 
 func (d *DataSet) Validate() error {